+++ test/ref_inspector/database/reload_test.exs
@@ -18,16 +18,14 @@ defmodule RefInspector.Database.ReloadTest do
 
   test "reloading databases" do
     Application.put_env(:ref_inspector, :database_files, [@fixture_search])
-    RefInspector.reload()
-    :timer.sleep(100)
+    RefInspector.reload(async: false)
 
     assert RefInspector.ready?()
     assert RefInspector.parse("http://www.google.com/test").source == "Google"
     assert RefInspector.parse("http://twitter.com/test").source == :unknown
 
     Application.put_env(:ref_inspector, :database_files, [@fixture_social])
-    RefInspector.reload()
-    :timer.sleep(100)
+    RefInspector.reload(async: false)
 
     assert RefInspector.ready?()
     assert RefInspector.parse("http://www.google.com/test").source == :unknown