+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Database entries are now stored in a single named table instead of using an intermediate reference table
     - Default configuration entries for files and urls are available through `RefInspector.Config.default_files/0` and `RefInspector.Config.default_urls/0`
     - Output of mix task `ref_inspector.download` can be prevented by passing `--quiet` upon invocation. This does NOT imply `--force` and will still ask for confirmation
+    - Passing `async: false` to `RefInspector.reload_databases/1` allows you to block your calling process until the reload has finished
     - The library used to download the database files can be changed by configuring a module implementing the `RefInspector.Downloader.Adapter` behaviour
     - The library used to read YAML files can be changed by using the `:yaml_file_reader` configuration